+++ spec/knapsack_pro/adapters/rspec_adapter_spec.rb
@@ -1,3 +1,5 @@
+require_relative '../../../lib/knapsack_pro/formatters/time_tracker'
+
 describe KnapsackPro::Adapters::RSpecAdapter do
   it 'backwards compatibility with knapsack gem old rspec adapter name' do
     expect(KnapsackPro::Adapters::RspecAdapter.new).to be_kind_of(described_class)
@@ -8,7 +10,7 @@ describe KnapsackPro::Adapters::RSpecAdapter do
   end
 
   context do
-    before { expect(::RSpec).to receive(:configure) }
+    before { expect(::RSpec).to receive(:configure).at_least(:once) }
     it_behaves_like 'adapter'
   end
 
@@ -237,59 +239,79 @@ describe KnapsackPro::Adapters::RSpecAdapter do
     end
   end
 
-  describe '.test_path' do
-    let(:example_group) do
-      {
-        file_path: '1_shared_example.rb',
-        parent_example_group: {
-          file_path: '2_shared_example.rb',
-          parent_example_group: {
-            file_path: 'a_spec.rb'
-          }
-        }
-      }
-    end
-    let(:current_example) do
-      OpenStruct.new(metadata: {
-        example_group: example_group
-      })
-    end
+  describe '.file_path_for' do
+    let(:current_example) { ::RSpec.describe.example }
 
-    subject { described_class.test_path(current_example) }
+    subject { described_class.file_path_for(current_example) }
 
-    it { should eql 'a_spec.rb' }
+    context "when id ends in _spec.rb" do
+      it "returns the first part of the id" do
+        allow(current_example).to receive(:id).and_return("./foo_spec.rb[1:1]")
 
-    context 'with turnip features' do
-      describe 'when the turnip version is less than 2' do
-        let(:example_group) do
-          {
-            file_path: "./spec/features/logging_in.feature",
-            turnip: true,
-            parent_example_group: {
-              file_path: "gems/turnip-1.2.4/lib/turnip/rspec.rb"
-            }
-          }
-        end
-
-        before { stub_const("Turnip::VERSION", '1.2.4') }
-
-        it { should eql './spec/features/logging_in.feature' }
+        expect(subject).to eq('./foo_spec.rb')
       end
+    end
 
-      describe 'when turnip is version 2 or greater' do
-        let(:example_group) do
-          {
-            file_path: "gems/turnip-2.0.0/lib/turnip/rspec.rb",
-            turnip: true,
+    context "when id does not end in _spec.rb" do
+      it "returns the file_path" do
+        allow(current_example).to receive(:id).and_return("./foo.rb")
+        allow(current_example).to receive(:metadata).and_return(file_path: "./foo_spec.rb")
+
+        expect(subject).to eq('./foo_spec.rb')
+      end
+    end
+
+    context "when id and file_path do not end in _spec.rb" do
+      it "returns the example_group's file_path" do
+        allow(current_example).to receive(:id).and_return("./foo.rb")
+        allow(current_example).to receive(:metadata).and_return(
+          file_path: "./foo.rb", example_group: { file_path: "./foo_spec.rb" }
+        )
+
+        expect(subject).to eq('./foo_spec.rb')
+      end
+    end
+
+    context "when id, file_path, and example_group's file_path do not end in _spec.rb" do
+      it "returns the top_level_group's file_path" do
+        allow(current_example).to receive(:id).and_return("./foo.rb")
+        allow(current_example).to receive(:metadata).and_return(
+          file_path: "./foo.rb",
+          example_group: {
+            file_path: "./foo.rb",
             parent_example_group: {
-              file_path: "./spec/features/logging_in.feature",
+              file_path: "./foo_spec.rb",
             }
           }
-        end
+        )
 
-        before { stub_const("Turnip::VERSION",  '2.0.0') }
+        expect(subject).to eq('./foo_spec.rb')
+      end
+    end
 
-        it { should eql './spec/features/logging_in.feature' }
+    context "when id, file_path, example_group's, and top_level_group's file_path do not end in _spec.rb" do
+      it "returns empty string" do
+        allow(current_example).to receive(:id).and_return("./foo.rb")
+        allow(current_example).to receive(:metadata).and_return(
+          file_path: "./foo.rb",
+          example_group: {
+            file_path: "./foo.rb",
+            parent_example_group: {
+              file_path: "./foo.rb",
+            }
+          }
+        )
+
+        expect(subject).to eq('')
+      end
+    end
+
+    context "when id does not end in .feature (nor _spec.rb)" do
+      it "returns the file_path" do
+        allow(current_example).to receive(:id).and_return("./foo.rb")
+        allow(current_example).to receive(:metadata).and_return(file_path: "./foo.feature")
+
+        expect(subject).to eq("./foo.feature")
       end
     end
   end
@@ -298,153 +320,46 @@ describe KnapsackPro::Adapters::RSpecAdapter do
     let(:config) { double }
 
     describe '#bind_time_tracker' do
-      let(:tracker) { instance_double(KnapsackPro::Tracker) }
-      let(:logger) { instance_double(Logger) }
-      let(:global_time) { 'Global time: 01m 05s' }
-      let(:test_path) { 'spec/a_spec.rb' }
       let(:current_example) { double(metadata: {}) }
 
       context "when the example's metadata has :focus tag AND RSpec inclusion rule includes :focus" do
         let(:current_example) { double(metadata: { focus: true }) }
+        let(:test_path) { 'spec/a_spec.rb' }
 
         it do
-          expect(KnapsackPro::Config::Env).to receive(:rspec_split_by_test_examples?).and_return(false)
-
-          expect(config).to receive(:prepend_before).with(:context).and_yield
-
-          allow(KnapsackPro).to receive(:tracker).and_return(tracker)
-          expect(tracker).to receive(:start_timer).ordered
-
           expect(config).to receive(:around).with(:each).and_yield(current_example)
           expect(::RSpec).to receive(:configure).and_yield(config)
 
-          expect(tracker).to receive(:current_test_path).ordered.and_return(test_path)
-          expect(tracker).to receive(:stop_timer).ordered
-
-          expect(described_class).to receive(:test_path).with(current_example).and_return(test_path)
-
-          expect(tracker).to receive(:current_test_path=).with(test_path).ordered
+          expect(described_class).to receive(:file_path_for).with(current_example).and_return(test_path)
 
           expect(described_class).to receive_message_chain(:rspec_configuration, :filter, :rules, :[]).with(:focus).and_return(true)
 
           expect {
             subject.bind_time_tracker
-          }.to raise_error /We detected a test file path spec\/a_spec\.rb with a test using the metadata `:focus` tag/
+          }.to raise_error /Knapsack Pro found an example tagged with focus in spec\/a_spec\.rb/i
         end
       end
 
-      context 'when rspec split by test examples is disabled' do
-        before do
-          expect(KnapsackPro::Config::Env).to receive(:rspec_split_by_test_examples?).and_return(false)
-        end
+      context 'with no focus' do
+        let(:logger) { instance_double(Logger) }
+        let(:duration) { 65 }
+        let(:global_time) { 'Global time execution for tests: 01m 05s' }
+        let(:time_tracker) { instance_double(KnapsackPro::Formatters::TimeTracker) }
 
         it 'records time for current test path' do
-          expect(config).to receive(:prepend_before).with(:context).and_yield
-
-          allow(KnapsackPro).to receive(:tracker).and_return(tracker)
-          expect(tracker).to receive(:start_timer).ordered
-
           expect(config).to receive(:around).with(:each).and_yield(current_example)
-          expect(config).to receive(:append_after).with(:context).and_yield
           expect(config).to receive(:after).with(:suite).and_yield
-          expect(::RSpec).to receive(:configure).and_yield(config)
-
-          expect(tracker).to receive(:current_test_path).ordered.and_return(test_path)
-          expect(tracker).to receive(:stop_timer).ordered
-
-          expect(described_class).to receive(:test_path).with(current_example).and_return(test_path)
-
-          expect(tracker).to receive(:current_test_path=).with(test_path).ordered
+          expect(::RSpec).to receive(:configure).twice.and_yield(config)
 
           expect(current_example).to receive(:run)
 
-          expect(tracker).to receive(:stop_timer).ordered
+          expect(time_tracker).to receive(:batch_duration).and_return(duration)
+          expect(KnapsackPro::Formatters::TimeTrackerFetcher).to receive(:call).and_return(time_tracker)
 
-          expect(KnapsackPro::Presenter).to receive(:global_time).and_return(global_time)
           expect(KnapsackPro).to receive(:logger).and_return(logger)
           expect(logger).to receive(:debug).with(global_time)
 
           subject.bind_time_tracker
-        end
-      end
-
-      context 'when rspec split by test examples is enabled' do
-        let(:test_example_path) { 'spec/a_spec.rb[1:1]' }
-
-        before do
-          expect(KnapsackPro::Config::Env).to receive(:rspec_split_by_test_examples?).and_return(true)
-        end
-
-        context 'when current test_path is a slow test file' do
-          before do
-            expect(described_class).to receive(:slow_test_file?).with(described_class, test_path).and_return(true)
-          end
-
-          it 'records time for example.id' do
-            expect(current_example).to receive(:id).and_return(test_example_path)
-
-            expect(config).to receive(:prepend_before).with(:context).and_yield
-
-            allow(KnapsackPro).to receive(:tracker).and_return(tracker)
-            expect(tracker).to receive(:start_timer).ordered
-
-            expect(config).to receive(:around).with(:each).and_yield(current_example)
-            expect(config).to receive(:append_after).with(:context).and_yield
-            expect(config).to receive(:after).with(:suite).and_yield
-            expect(::RSpec).to receive(:configure).and_yield(config)
-
-            expect(tracker).to receive(:current_test_path).ordered.and_return(test_path)
-            expect(tracker).to receive(:stop_timer).ordered
-
-            expect(described_class).to receive(:test_path).with(current_example).and_return(test_path)
-
-            expect(tracker).to receive(:current_test_path=).with(test_example_path).ordered
-
-            expect(current_example).to receive(:run)
-
-            expect(tracker).to receive(:stop_timer).ordered
-
-            expect(KnapsackPro::Presenter).to receive(:global_time).and_return(global_time)
-            expect(KnapsackPro).to receive(:logger).and_return(logger)
-            expect(logger).to receive(:debug).with(global_time)
-
-            subject.bind_time_tracker
-          end
-        end
-
-        context 'when current test_path is not a slow test file' do
-          before do
-            expect(described_class).to receive(:slow_test_file?).with(described_class, test_path).and_return(false)
-          end
-
-          it 'records time for current test path' do
-            expect(config).to receive(:prepend_before).with(:context).and_yield
-
-            allow(KnapsackPro).to receive(:tracker).and_return(tracker)
-            expect(tracker).to receive(:start_timer).ordered
-
-            expect(config).to receive(:around).with(:each).and_yield(current_example)
-            expect(config).to receive(:append_after).with(:context).and_yield
-            expect(config).to receive(:after).with(:suite).and_yield
-            expect(::RSpec).to receive(:configure).and_yield(config)
-
-            expect(described_class).to receive(:test_path).with(current_example).and_return(test_path)
-
-            expect(tracker).to receive(:current_test_path).ordered.and_return(test_path)
-            expect(tracker).to receive(:stop_timer).ordered
-
-            expect(tracker).to receive(:current_test_path=).with(test_path).ordered
-
-            expect(current_example).to receive(:run)
-
-            expect(tracker).to receive(:stop_timer).ordered
-
-            expect(KnapsackPro::Presenter).to receive(:global_time).and_return(global_time)
-            expect(KnapsackPro).to receive(:logger).and_return(logger)
-            expect(logger).to receive(:debug).with(global_time)
-
-            subject.bind_time_tracker
-          end
         end
       end
     end
@@ -454,7 +369,11 @@ describe KnapsackPro::Adapters::RSpecAdapter do
         expect(config).to receive(:after).with(:suite).and_yield
         expect(::RSpec).to receive(:configure).and_yield(config)
 
-        expect(KnapsackPro::Report).to receive(:save)
+        time_tracker = instance_double(KnapsackPro::Formatters::TimeTracker)
+        times = [{ path: "foo_spec.rb", time_execution: 1.0 }]
+        expect(time_tracker).to receive(:batch).and_return(times)
+        expect(KnapsackPro::Formatters::TimeTrackerFetcher).to receive(:call).and_return(time_tracker)
+        expect(KnapsackPro::Report).to receive(:save).with(times)
 
         subject.bind_save_report
       end