+++ spec/knapsack_pro/extensions/time_spec.rb
@@ -1,5 +1,0 @@
-describe Time do
-  it do
-    expect(Time.respond_to?(:raw_now)).to be true
-  end
-end